feat: Add plug method to BklarApp and Router
Introduces a new `plug` method to the `BklarApp` and `Router` classes. This allows for the integration of routes from other Bklar applications (sub-apps) into the main application's routing configuration. This promotes modularity and reusability of route definitions. The version of the bklar package has also been updated to 1.0.6 to reflect these changes.