feat: Add createdAt and date fields to schemas and update swagger
Adds `createdAt` to the `userSchema` and `date` to the `loginSchema` in the `example-api`.
Also, updates the `@bklarjs/swagger` package to version `1.1.1` and enhances its OpenAPI generation to correctly map ZodDate types to "string" with "date-time" format. This ensures proper representation of date fields in generated OpenAPI documentation. Updates dependencies in `bun.lock`.